requirements.txt
antlr4-python3-runtime==4.9.3
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_3beajm7umk/croot/certifi_1745939228545/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cloudpickle==3.1.1
cn-clip==1.5.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fire==0.7.0
fonttools==4.58.4
fsspec==2025.5.1
fvcore==0.1.5.post20221221
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.32.5
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.3.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.4
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
portalocker==3.2.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.2.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tabulate==0.9.0
termcolor==3.1.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.2
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.67.1
transformers==4.52.4
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
urllib3==2.4.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
yacs==0.1.8
zipp==3.23.0
